requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
boto3==1.40.55
botocore==1.40.55
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
jmespath==1.0.1
psycopg==3.2.11
psycopg-binary==3.2.11
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
